trip-itinerary-service/pyproject.toml
[project]
name = "itinerary"
version = "0.1.0"
readme = "README.md"
authors = [
    { name = "Senyu Xie", email = "senyuxie@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.75.0",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "tripsphere",
]

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "grpcio-tools>=1.75.0",
    "mypy>=1.18.2",
    "mypy-protobuf>=3.6.0",
    "types-grpcio>=1.0.0.20250914",
]

[tool.uv.sources]
tripsphere = { workspace = true }

[tool.uv.workspace]
members = ["libs/*"]

[tool.ruff]
extend = "../ruff.toml"
extend-exclude = ["libs/tripsphere"]
